+++ packages/tizen_audio_manager/example/integration_test/audio_manager_test.dart
@@ -60,12 +60,12 @@ void main() {
   testWidgets('test alarm set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.alarm);
-    AudioManager.volumeController.setLevel(AudioVolumeType.alarm, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.alarm, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.alarm);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.alarm, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.alarm, 0);
     level = await AudioManager.volumeController.getLevel(AudioVolumeType.alarm);
     expect(level, equals(0));
   });
@@ -73,12 +73,12 @@ void main() {
   testWidgets('test call set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.call);
-    AudioManager.volumeController.setLevel(AudioVolumeType.call, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.call, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.call);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.call, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.call, 0);
     level = await AudioManager.volumeController.getLevel(AudioVolumeType.call);
     expect(level, equals(0));
   });
@@ -86,12 +86,12 @@ void main() {
   testWidgets('test media set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.media);
-    AudioManager.volumeController.setLevel(AudioVolumeType.media, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.media, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.media);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.media, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.media, 0);
     level = await AudioManager.volumeController.getLevel(AudioVolumeType.media);
     expect(level, equals(0));
   });
@@ -99,12 +99,14 @@ void main() {
   testWidgets('test notification set level', (WidgetTester tester) async {
     final int max = await AudioManager.volumeController
         .getMaxLevel(AudioVolumeType.notification);
-    AudioManager.volumeController.setLevel(AudioVolumeType.notification, max);
+    await AudioManager.volumeController
+        .setLevel(AudioVolumeType.notification, max);
     int level = await AudioManager.volumeController
         .getLevel(AudioVolumeType.notification);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.notification, 0);
+    await AudioManager.volumeController
+        .setLevel(AudioVolumeType.notification, 0);
     level = await AudioManager.volumeController
         .getLevel(AudioVolumeType.notification);
     expect(level, equals(0));
@@ -113,12 +115,12 @@ void main() {
   testWidgets('test ringtone set level', (WidgetTester tester) async {
     final int max = await AudioManager.volumeController
         .getMaxLevel(AudioVolumeType.ringtone);
-    AudioManager.volumeController.setLevel(AudioVolumeType.ringtone, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.ringtone, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.ringtone);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.ringtone, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.ringtone, 0);
     level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.ringtone);
     expect(level, equals(0));
@@ -127,12 +129,12 @@ void main() {
   testWidgets('test system set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.system);
-    AudioManager.volumeController.setLevel(AudioVolumeType.system, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.system, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.system);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.system, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.system, 0);
     level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.system);
     expect(level, equals(0));
@@ -141,12 +143,12 @@ void main() {
   testWidgets('test voice set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.voice);
-    AudioManager.volumeController.setLevel(AudioVolumeType.voice, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.voice, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.voice);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.voice, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.voice, 0);
     level = await AudioManager.volumeController.getLevel(AudioVolumeType.voice);
     expect(level, equals(0));
   });
@@ -154,12 +156,12 @@ void main() {
   testWidgets('test voip set level', (WidgetTester tester) async {
     final int max =
         await AudioManager.volumeController.getMaxLevel(AudioVolumeType.voip);
-    AudioManager.volumeController.setLevel(AudioVolumeType.voip, max);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.voip, max);
     int level =
         await AudioManager.volumeController.getLevel(AudioVolumeType.voip);
     expect(level, equals(max));
 
-    AudioManager.volumeController.setLevel(AudioVolumeType.voip, 0);
+    await AudioManager.volumeController.setLevel(AudioVolumeType.voip, 0);
     level = await AudioManager.volumeController.getLevel(AudioVolumeType.voip);
     expect(level, equals(0));
   });