+++ packages/webview_flutter_lwe/lib/webview_flutter_lwe.dart
@@ -3,8 +3,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-library webview_flutter_lwe;
-
 export 'src/lwe_webview_controller.dart';
 export 'src/lwe_webview_cookie_manager.dart';
 export 'src/lwe_webview_platform.dart';