+++ packages/flutter_app_badger/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## NEXT
 
-* Increase the minimum Flutter version to 3.3.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
 
 ## 0.1.0