+++ packages/flutter_secure_storage/CHANGELOG.md
@@ -1,3 +1,8 @@
+## NEXT
+
+* Fix new lint warnings.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 0.1.1
 
 * Update flutter_secure_storage to 9.0.0.