+++ packages/flutter_webrtc/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 0.1.3
 
 * Increase the minimum Flutter version to 3.3.