+++ packages/geolocator/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 1.0.6
 
 * Remove unnecessary `StreamHandlerError` implementation.