+++ packages/integration_test/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
+
 ## 2.0.2
 
 * Increase the minimum Flutter version to 3.3.