+++ packages/tizen_rpc_port/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## NEXT
 
 * Fix token in finalizer to be different from value.
+* Fix new lint warnings.
+* Update minimum Flutter and Dart version to 3.13 and 3.1.
 
 ## 0.1.3